Parallelize S3 uploads
With the Snacks dataset (848 MB, 6745 images) for example, this change
brings the upload time down from 12 minutes to 1 minute.

Note: if we set MAX_PARALLEL_UPLOADS to 100, we run into some issues
with S3 rate limiting. 32 seems to work with no issues, but in the
future, this parameter should probably be chosen more intelligently.