Add basic retry logic for S3 uploads
This patch adds retries with exponential backoff to S3 uploads. The
implementation in this patch isn't ideal, because it only makes the
failed upload back off, but all other coroutines (up to 31, with the
current choice of parameters) continue uploading as before. Still, it
seems to mostly alleviate the problem with failed S3 uploads due to rate
limiting or transient errors.