Fix variables showing up in docs as globals
This patch fixes "name" and "method" showing up in the list of global
variables defined by this module (which was being rendered in the docs).

There are some other preexisting problems with the telemetry, including
missing telemetry for the Model and TLM classes; this patch does not
address those issues.